+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## v1.2.9
 
- 
+### Added
+
+- command `depreset` to reset the status of a given deployment
+
 ### Changed
 
 - command `depshow` and `depcreate` now support option `--ojson` that formats the command output in json